requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
python-multipart>=0.0.20
pydantic>=2.8.0
requests>=2.32.0
python-dotenv>=1.0.1
openai-whisper>=20250625
numpy>=1.26.0
